[IMP] delivery_roulier_picking_batch: Handle multi packages from multiple pickings
