+++ delivery_roulier_picking_batch/models/stock_picking.py
@@ -12,19 +12,21 @@ class StockPicking(models.Model):
 
     def _is_batch_roulier(self):
         # Check if the picking is part of a batch with a roulier carrier
-        self.ensure_one()
+
         return (
-            self.batch_id
+            len(self.mapped("batch_id")) == 1
+            and self.batch_id
             and self.batch_id.carrier_id
             and self.batch_id.carrier_id._is_roulier()
         )
 
     def send_to_shipper(self):
-        self.ensure_one()
         if self._is_batch_roulier():
             # We are in a batch with a roulier carrier
             # We need to send unsent packages
             packages = self.package_ids - self.batch_id.sent_package_ids
+            # Do not send packages that already have a tracking number
+            packages = packages.filtered(lambda p: not p.parcel_tracking)
             if not packages:
                 # Nothing to send
                 return
@@ -98,13 +100,24 @@ class StockPicking(models.Model):
 
         return super().send_to_shipper()
 
+    def _send_confirmation_email(self):
+        # Bypass odoo's foreach picking that prevents to call a single
+        # send_to_shipper with all batch packages independently of the pickings
+        if self._is_batch_roulier():
+            self.sudo().send_to_shipper()
+
+        # Call the super method to handle the email sending
+        return super()._send_confirmation_email()
+
     def _roulier_generate_labels(self):
         if self._is_batch_roulier():
             label_info = []
-            for picking in self:
-                # Generate labels only for unsent packages
-                packages = picking.package_ids - picking.batch_id.sent_package_ids
-                label_info.append(packages._generate_labels(picking))
+
+            # Generate labels only for unsent packages
+            packages = self.package_ids - self.batch_id.sent_package_ids
+            # Do not send packages that already have a tracking number
+            packages = packages.filtered(lambda p: not p.parcel_tracking)
+            label_info.append(packages._generate_labels(self[0]))
             return label_info
 
         return super()._roulier_generate_labels()
